# Draft: Optimise walkpath