Closes #158

I think this is what we want. I've changed the original suggestion to return a `Channel` rather than a vector to align with existing behaviour. The logic for inserting the paths into the channel when `topdown = false / true` is also bit fiddly but I hope it makes sense. I can add more comments if folks like.

We can likely improve on this even more so feedback welcome.

for comparison on the speedup on a non-trivial example
```julia

# current
julia> @time collect(walkpath(path10; topdown=true));
 24.777315 seconds (164.21 k allocations: 10.317 MiB)

julia> @time collect(walkpath(path10; topdown=false));
 28.085743 seconds (164.48 k allocations: 10.400 MiB)

# this PR
julia> @time collect(walkpath(path10; topdown=true));
  0.224652 seconds (10.23 k allocations: 563.859 KiB)

julia> @time collect(walkpath(path10; topdown=false));
  0.211930 seconds (10.31 k allocations: 572.734 KiB)
```