@glennmoy 2021-08-03
2 tests from `test_walkpath` seem to be failing. Not sure what the issue is yet.

```julia
julia> result
9-element Vector{S3Path{AWS.AWSConfig}}:
 p"s3://ocaws.jl.test.20210803t192655z/pathset-root/./"
 p"s3://ocaws.jl.test.20210803t192655z/pathset-root"
 p"s3://ocaws.jl.test.20210803t192655z/pathset-root/bar/"
 p"s3://ocaws.jl.test.20210803t192655z/pathset-root/bar/qux/"
 p"s3://ocaws.jl.test.20210803t192655z/pathset-root/bar/qux/quux.tar.gz"
 p"s3://ocaws.jl.test.20210803t192655z/pathset-root/foo/"
 p"s3://ocaws.jl.test.20210803t192655z/pathset-root/foo/baz.txt"
 p"s3://ocaws.jl.test.20210803t192655z/pathset-root/fred/"
 p"s3://ocaws.jl.test.20210803t192655z/pathset-root/fred/plugh"

julia> expected
7-element Vector{S3Path{AWS.AWSConfig}}:
 p"s3://ocaws.jl.test.20210803t192655z/pathset-root/bar/"
 p"s3://ocaws.jl.test.20210803t192655z/pathset-root/bar/qux/"
 p"s3://ocaws.jl.test.20210803t192655z/pathset-root/bar/qux/quux.tar.gz"
 p"s3://ocaws.jl.test.20210803t192655z/pathset-root/foo/"
 p"s3://ocaws.jl.test.20210803t192655z/pathset-root/foo/baz.txt"
 p"s3://ocaws.jl.test.20210803t192655z/pathset-root/fred/"
 p"s3://ocaws.jl.test.20210803t192655z/pathset-root/fred/plugh"
```