@rofinn 2021-08-04
Yeah, I noticed that when working on an implementation that doesn't require collecting the results. FWIW, I think #179 fixes the issue and only requires a single iteration over the `s3_list_objects` results.

```julia
# The current master branch NA... gave up after 10 min :)

# gm/walkpath
julia> @time collect(walkpath(fp; topdown=false));
 46.573418 seconds (285.25 M allocations: 12.779 GiB, 3.07% gc time)

julia> @time collect(walkpath(fp; topdown=true));
 45.402200 seconds (285.14 M allocations: 12.773 GiB, 3.40% gc time)

# rf/walkpath
julia> @time collect(walkpath(fp; topdown=false));
 17.915066 seconds (1.10 M allocations: 70.752 MiB, 0.21% gc time)

julia> @time collect(walkpath(fp; topdown=true));
 19.116504 seconds (1.10 M allocations: 70.328 MiB, 0.10% gc time)
```